+++ time/src/main/java/net/jqwik/time/api/Dates.java
@@ -5,7 +5,8 @@ import java.time.*;
 import org.apiguardian.api.*;
 
 import net.jqwik.api.*;
-import net.jqwik.time.*;
+import net.jqwik.time.api.arbitraries.*;
+import net.jqwik.time.engine.properties.arbitraries.*;
 
 import static org.apiguardian.api.API.Status.*;
