+++ time/src/test/java/net/jqwik/time/api/DatesTests.java
@@ -7,6 +7,8 @@ import net.jqwik.api.*;
 import net.jqwik.api.constraints.*;
 import net.jqwik.api.statistics.*;
 import net.jqwik.testing.*;
+import net.jqwik.time.api.annotations.*;
+import net.jqwik.time.api.arbitraries.*;
 
 import static org.assertj.core.api.Assertions.*;
